+++ src/lib/axios/index.spec.ts
@@ -1,0 +1,7 @@
+import axios from '.';
+
+describe('axios import test', function () {
+  it('can import the module', function () {
+    expect(axios).toBeInstanceOf(Function);
+  });
+});