+++ src/lib/fast-json-stable-stringify/index.spec.ts
@@ -1,0 +1,7 @@
+import { stringify } from '.';
+
+describe('fast-json-stable-stringify import test', function () {
+  it('can import the module', function () {
+    expect(stringify).toBeInstanceOf(Function);
+  });
+});