+++ src/lib/fast-json-stable-stringify/index.ts
@@ -1,0 +1,2 @@
+import stringify from 'fast-json-stable-stringify';
+export { stringify };