+++ src/lib/joi/index.spec.ts
@@ -1,0 +1,10 @@
+import Joi from '.';
+
+describe('joi import test', function () {
+  it('can import the module', function () {
+    expect(Joi).toBeInstanceOf(Object);
+    expect(Joi.boolean).toBeInstanceOf(Function);
+    expect(Joi.number).toBeInstanceOf(Function);
+    expect(Joi.string).toBeInstanceOf(Function);
+  });
+});