+++ src/lib/json5/index.spec.ts
@@ -1,0 +1,8 @@
+import { parse, stringify } from '.';
+
+describe('json5 import test', function () {
+  it('can import the module', function () {
+    expect(parse).toBeInstanceOf(Function);
+    expect(stringify).toBeInstanceOf(Function);
+  });
+});