+++ src/lib/json5/index.ts
@@ -1,0 +1,3 @@
+import json5 from 'json5';
+export default json5;
+export * from 'json5';