+++ CHANGELOG.md
@@ -1,3 +1,23 @@
+## 1.1.0
+
+### Dependencies
+
+- `[@types/lodash]` N/A ➞ 4.14.175
+- `[@types/luxon]` N/A ➞ 2.0.5
+- `[@types/prettier]` N/A ➞ 2.4.1
+- `[@types/yargs]` N/A ➞ 17.0.4
+- `[axios]` N/A ➞ 0.23.0
+- `[chalk]` N/A ➞ 4.1.2
+- `[fast-json-stable-stringify]` N/A ➞ 2.1.0
+- `[http-status-codes]` N/A ➞ 2.1.4
+- `[joi]` N/A ➞ 17.4.2
+- `[json5]` N/A ➞ 2.2.0
+- `[lodash]` N/A ➞ 4.17.21
+- `[luxon]` N/A ➞ 2.0.2
+- `[prettier]` N/A ➞ 2.4.1
+- `[tslib]` N/A ➞ 2.3.1
+- `[yargs]` N/A ➞ 17.2.1
+
 ## 1.0.0
 
 - Initial version